+++ src/lib/viewers/box3d/image360/__tests__/Image360Viewer-test.js
@@ -3,6 +3,7 @@ import Image360Viewer from '../Image360Viewer';
 import BaseViewer from '../../../BaseViewer';
 import Box3DControls from '../../Box3DControls';
 import Image360Renderer from '../Image360Renderer';
+import { SELECTOR_BOX_PREVIEW_CONTENT } from '../../../../constants';
 
 const sandbox = sinon.sandbox.create();
 const CSS_CLASS_IMAGE_360 = 'bp-image-360';
@@ -27,7 +28,7 @@ describe('lib/viewers/box3d/image360/Image360Viewer', () => {
             }
         });
         Object.defineProperty(BaseViewer.prototype, 'setup', { value: sandbox.mock() });
-        viewer.containerEl = containerEl;
+        viewer.containerEl = document.querySelector(SELECTOR_BOX_PREVIEW_CONTENT);
     });
 
     afterEach(() => {